+++ sdmx-dl-api/src/main/java/sdmxdl/Dataflow.java
@@ -19,7 +19,8 @@ package sdmxdl;
 /**
  * @author Philippe Charles
  */
-@lombok.Value(staticConstructor = "of")
+@lombok.Value
+@lombok.Builder(toBuilder = true)
 @lombok.EqualsAndHashCode(callSuper = false)
 public class Dataflow extends Resource<DataflowRef> {