+++ sdmx-dl-format-xml/src/test/java/internal/sdmxdl/format/xml/XMLStreamFlow20Test.java
@@ -41,8 +41,18 @@ public class XMLStreamFlow20Test {
 
         assertThat(p.parseReader(SdmxXmlSources.OTHER_FLOWS20::openReader))
                 .containsExactly(
-                        Dataflow.of(DataflowRef.of("IMF", "DS-BOP_2017M06", "1.0"), DataStructureRef.of("IMF", "BOP_2017M06", null), "Balance of Payments (BOP), 2017 M06"),
-                        Dataflow.of(DataflowRef.of("IMF", "DS-BOP_2020M3", "1.0"), DataStructureRef.of("IMF", "BOP_2020M3", null), "Balance of Payments (BOP), 2020 M03")
+                        Dataflow
+                                .builder()
+                                .ref(DataflowRef.of("IMF", "DS-BOP_2017M06", "1.0"))
+                                .structureRef(DataStructureRef.of("IMF", "BOP_2017M06", null))
+                                .label("Balance of Payments (BOP), 2017 M06")
+                                .build(),
+                        Dataflow
+                                .builder()
+                                .ref(DataflowRef.of("IMF", "DS-BOP_2020M3", "1.0"))
+                                .structureRef(DataStructureRef.of("IMF", "BOP_2020M3", null))
+                                .label("Balance of Payments (BOP), 2020 M03")
+                                .build()
                 );
     }
 }